// server/types.ts
export interface Answer {
  _id?: ObjectId;
  text: string;
  ansBy: string;
  ansDateTime: Date;
  comments: Comment[];
}
// ...
export interface Tag {
  _id?: ObjectId;
  name: string;
  description: string;
}
// ...
export interface Question {
  _id?: ObjectId;
  title: string;
  text: string;
  tags: Tag[];
  askedBy: string;
  askDateTime: Date;
  answers: Answer[];
  views: number;
  upVotes: string[];
  downVotes: string[];
  comments: Comment[];

}
// ...
export interface Comment {
  _id?: ObjectId;
  text: string;
  commentBy: string;
  commentDateTime: Date;
}